Moving compat class
